✨ feat: 커스텀 훅으로 리팩토링 이전 컴포넌트와 동일하게 로직분리와 UI 를 나누어서 재사용성을 높히고자 커스텀훅으로 로직을 옮겼습니다. Resolve: #153
